requirement.txt
requests
argparse
pycryptodome